requirements.txt
google-adk==1.19.0
google-cloud-aiplatform[agent_engines,adk]>=1.112
python-dotenv>=1.0.0